requirements.txt
python-telegram-bot>=21.0
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
chromadb>=0.4.22
sqlalchemy>=2.0.25
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
aiosqlite>=0.19.0
